feat: 입력에 대한 예외 처리만을 위한 클래스 생성 후 분리
input은 BufferedReader 변수를 넘겨주기만 하고, Validator 클래스에서 이를 관리한다.
